profile pictures are now directly uploaded to cloudinary

+++ controllers/userController.js
@@ -110,20 +110,16 @@ const registerUser   = async (req, res) => {
       });
 
       // Clean up the uploaded file if it exists
-           // Clean up the uploaded file if it exists
-      /* if (profilePicture && fs.existsSync(profilePicture.path)) {
-          try {
-              fs.unlinkSync(profilePicture.path); // Delete the file only if it exists
-          } catch (err) {
-              console.error('Error deleting file:', err); // Log any errors during deletion
-          }
-      } */ // Close the if block
-
+      if (profilePicture && fs.existsSync(profilePicture.path)) {
+          fs.unlinkSync(profilePicture.path);
+      }
   } catch (err) {
-      console.error(err);
+      console.log(err);
       return res.status(500).json({ message: err.message });
-  } // Close the try-catch block
-}; // Close the registerUser function
+  }
+};
+
+
 //The following function is for signing in a user to the vaarbz application
